requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlmodel==0.0.24
SQLAlchemy==2.0.40
alembic==1.11.1
APScheduler==3.11.1
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.9.1
dnspython==2.7.0

# Optional / development
# httpx can be used by clients if needed
# httpx==0.28.1
